auditlog: refactor ydb/core/audit library interface
- separate interface into client and service sides
- change CreateAuditWriter logBackends arg to rvalue ref

KIKIMR-21774
